+++ CHANGES.rst
@@ -7,7 +7,7 @@ Changelog
 - added `set-placeholder` filter for dynamically setting (temporary) placeholders at runtime
 - added `--resume_from` option to relevant readers that allows resuming the data processing
   from the first file that matches this glob expression (e.g., `*/012345.wav`)
-- requiring seppl>=0.2.14 now for resume support
+- requiring seppl>=0.2.15 now for resume and split group support
 
 
 0.0.2 (2025-03-14)